requirements.txt
pandas 
numpy
scikit-learn
python-dotenv
dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.6
scikit-learn
PyYAML
dill


-e .
